Fix using int instead of size_t for strcspn return

+++ libsoup/soup-headers.c
@@ -907,7 +907,7 @@ append_param_quoted (GString    *string,
 		     const char *name,
 		     const char *value)
 {
-	int len;
+	gsize len;
 
 	g_string_append (string, name);
 	g_string_append (string, "=\"");
